Move tweens to AnimationSystem
