+++ packages/engine/src/avatar/AnimationSystem.ts
@@ -1,4 +1,4 @@
-import { Vector3 } from 'three'
+import { Euler, Vector3 } from 'three'
 import matches from 'ts-matches'
 
 import { Engine } from '../ecs/classes/Engine'
@@ -6,6 +6,9 @@ import { World } from '../ecs/classes/World'
 import { defineQuery, getComponent } from '../ecs/functions/ComponentFunctions'
 import { NetworkObjectComponent } from '../networking/components/NetworkObjectComponent'
 import { NetworkWorldAction } from '../networking/functions/NetworkWorldAction'
+import { DesiredTransformComponent } from '../transform/components/DesiredTransformComponent'
+import { TransformComponent } from '../transform/components/TransformComponent'
+import { TweenComponent } from '../transform/components/TweenComponent'
 import { AnimationManager } from './AnimationManager'
 import { AnimationGraph } from './animations/AnimationGraph'
 import { AnimationRenderer } from './animations/AnimationRenderer'
@@ -14,6 +17,13 @@ import { AvatarStates } from './animations/Util'
 import { AnimationComponent } from './components/AnimationComponent'
 import { AvatarAnimationComponent } from './components/AvatarAnimationComponent'
 
+const euler1YXZ = new Euler()
+euler1YXZ.order = 'YXZ'
+const euler2YXZ = new Euler()
+euler2YXZ.order = 'YXZ'
+
+const desiredTransformQuery = defineQuery([DesiredTransformComponent])
+const tweenQuery = defineQuery([TweenComponent])
 const animationQuery = defineQuery([AnimationComponent])
 const avatarAnimationQuery = defineQuery([AnimationComponent, AvatarAnimationComponent])
 
@@ -40,6 +50,36 @@ export default async function AnimationSystem(world: World) {
 
   return () => {
     const { delta } = world
+
+    for (const entity of desiredTransformQuery(world)) {
+      const desiredTransform = getComponent(entity, DesiredTransformComponent)
+
+      const mutableTransform = getComponent(entity, TransformComponent)
+      mutableTransform.position.lerp(desiredTransform.position, desiredTransform.positionRate * delta)
+
+      // store rotation before interpolation
+      euler1YXZ.setFromQuaternion(mutableTransform.rotation)
+      // lerp to desired rotation
+
+      mutableTransform.rotation.slerp(desiredTransform.rotation, desiredTransform.rotationRate * delta)
+      euler2YXZ.setFromQuaternion(mutableTransform.rotation)
+      // use axis locks - yes this is correct, the axis order is weird because quaternions
+      if (desiredTransform.lockRotationAxis[0]) {
+        euler2YXZ.x = euler1YXZ.x
+      }
+      if (desiredTransform.lockRotationAxis[2]) {
+        euler2YXZ.y = euler1YXZ.y
+      }
+      if (desiredTransform.lockRotationAxis[1]) {
+        euler2YXZ.z = euler1YXZ.z
+      }
+      mutableTransform.rotation.setFromEuler(euler2YXZ)
+    }
+
+    for (const entity of tweenQuery(world)) {
+      const tween = getComponent(entity, TweenComponent)
+      tween.tween.update()
+    }
 
     for (const entity of animationQuery(world)) {
       const animationComponent = getComponent(entity, AnimationComponent)
